adicionada a validação de id nas rotas

+++ routes/foodRoutes.js
@@ -16,6 +16,9 @@ router.get('/:id', async(req,res)=>{
     const {id} = req.params
     try {
         const food = await Food.findById(id)
+        if(!food)
+            return res.status(400).send('Id inválido!')
+            
         res.send({food})
     } catch (error) {
         res.status(400).send({error:"Não foi possível buscar o lanche "+error})
@@ -25,9 +28,9 @@ router.get('/:id', async(req,res)=>{
 router.post('/',async(req,res)=>{
     const {name,link_picture,price,description} = req.body
     try {
-        const food = await Food.find({name})
-        if(food)
-            res.status(400).send("Você ja cadastrou um lanche com esse nome!!")
+    
+        if(await Food.findOne({name}))
+           return res.status(400).send("Você ja cadastrou um lanche com esse nome!!")
             
         const newFood = await Food.create({name,link_picture,price,description})
 
@@ -42,6 +45,8 @@ router.put('/:id', async(req,res)=>{
     const {name, link_picture,price,description} = req.body
     const {id} = req.params
     try {
+        if(!await Food.findById(id) )
+            return res.status(400).send('Id inválido!')
 
         const food = await Food.findByIdAndUpdate(id,{name,link_picture,price,description}, {new:true})
 
@@ -56,10 +61,13 @@ router.put('/:id', async(req,res)=>{
 router.delete('/:id', async(req,res)=>{
     const {id} = req.params
     try {
-        const user = await Food.findByIdAndDelete(id)
+        if(!await Food.findById(id) )
+            return res.status(400).send('Id inválido!')
+
+        await Food.findByIdAndDelete(id)
         res.send({ok:"Lanche removido com sucesso!"})
     } catch (error) {
-        res.status(400).send({error:"Nçao foi possível remover o lanche! "+error})
+        res.status(400).send({error:"Não foi possível remover o lanche! "+error})
     }
 })
